requirements.txt
aiogram==3.6.0
python-dotenv==1.0.0
peewee==3.17.5
psycopg2==2.9.9
google-api-python-client==2.129.0
oauth2client==4.1.3
pygsheets==2.0.6
pytz==2024.1